concat test files
